refactor: 자동차의 이름 객체로 분리
