+++ src/main/java/domain/Car.java
@@ -1,14 +1,11 @@
 package domain;
 
-import org.apache.commons.lang3.StringUtils;
-
 public class Car {
     public static final int START_POSITION = 0;
-    private static final int MAX_NAME_LENGTH = 5;
     private static final int MOVE_NUMBER = 4;
     private static final int PER_MOVEMENT_RANGE = 1;
 
-    private final String name;
+    private final Name name;
     private int position;
 
     public Car(final String name) {
@@ -16,32 +13,18 @@ public class Car {
     }
 
     public Car(final String name, final int position) {
-        validateName(name);
+        this(new Name(name), position);
+    }
+
+    public Car(final Name name, final int position) {
         validatePosition(position);
         this.name = name;
         this.position = position;
     }
 
-    private void validateName(final String name) {
-        validarteNullOrEmpty(name);
-        validateNameLength(name);
-    }
-
-    private void validatePosition(int position) {
+    private void validatePosition(final int position) {
         if (position < START_POSITION) {
             throw new IllegalArgumentException("위치 값은 0 이상이여야 합니다.");
-        }
-    }
-
-    private void validateNameLength(final String name) {
-        if (name.length() > MAX_NAME_LENGTH) {
-            throw new IllegalArgumentException("이름은 5자 이하여야 합니다.");
-        }
-    }
-
-    private void validarteNullOrEmpty(final String name) {
-        if (StringUtils.isBlank(name)) {
-            throw new IllegalArgumentException("한 자리 이상의 이름을 입력해주세요.");
         }
     }
 
@@ -50,7 +33,7 @@ public class Car {
     }
 
     public String name() {
-        return name;
+        return name.value();
     }
 
     public void move(final int randomNumber) {
