+++ src/test/java/domain/CarTest.java
@@ -4,7 +4,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.NullAndEmptySource;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
@@ -35,26 +34,6 @@ public class CarTest {
         //when //then
         assertThatIllegalArgumentException().isThrownBy(() -> new Car(carName, negativePosition))
                 .withMessage("위치 값은 0 이상이여야 합니다.");
-    }
-
-    @ParameterizedTest
-    @NullAndEmptySource
-    @DisplayName("null 또는 공백일 경우, 예외가 발생한다,")
-    void validateNullOrEmpty(String nullOrEmpty) {
-        //given //when //then
-        assertThatIllegalArgumentException().isThrownBy(() -> new Car(nullOrEmpty))
-                .withMessage("한 자리 이상의 이름을 입력해주세요.");
-    }
-
-    @Test
-    @DisplayName("이름이 5자 초과일 경우, 예외가 발생한다.")
-    void validateNameLength() {
-        //given
-        String overLengthName = "abcdef";
-
-        //when //then
-        assertThatIllegalArgumentException().isThrownBy(() -> new Car(overLengthName))
-                .withMessage("이름은 5자 이하여야 합니다.");
     }
 
     @ParameterizedTest
